+++ lib/src/text/str_lookup.dart
@@ -44,6 +44,9 @@ class MapStrLookup extends StrLookup {
   /// Create a new instance with the given map.
   MapStrLookup({required this.map});
 
+  /// Looks up a String key to a String value using the map.
+  /// If the map is null, then null is returned.
+  /// The map result object is converted to a string using toString().
   @override
   String? lookup(String? key) {
     if (map == null) {