+++ test/text/str_matcher_test.dart
@@ -1,9 +1,9 @@
 // Licensed to the Limeslice Software Foundation (LSF) under one or more
 // contributor license agreements.  See the NOTICE file distributed with
 // this work for additional information regarding copyright ownership.
-// The LSF licenses this file to You under the MIT License (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
+// The LSF licenses this file to You under the Apache License, Version 2.0
+// (the "License"); you may not use this file except in compliance with
+// the License.  You may obtain a copy of the License at
 //
 // https://limeslice.org/license.txt
 //
@@ -13,11 +13,15 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-/// Commons Lang provides a host of helper utilities for Dart / Flutter
-/// applications, most notably String manipulation methods and basic numerical
-/// methods.
-library;
+import 'package:commons_lang/commons_lang.dart';
+import 'package:test/test.dart';
 
-export 'src/bool_utils.dart';
-export 'src/text/str_lookup.dart';
-export 'src/text/str_matcher.dart';
+void main() {
+  test('Test String Matcher', () {
+    String buffer = "abcdef";
+    StrMatcher matcher = StrMatcher.stringMatcher("bc");
+    expect(matcher.isMatch(buffer, 0), equals(0));
+    expect(matcher.isMatch(buffer, 1), equals(2));
+    expect(matcher.isMatch(buffer, 2), equals(0));
+  });
+}