+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## 0.0.3
 - Add StrLookup class.
+- Add StrMatcher class.
 
 ## 0.0.2
 - Add BoolUtils class.